Escape special characters in filetypes
Filetypes are treated as parts of variable names (using eval()). However, some
characters are legal in a filetype but not in an identifier. This caused the
Neomake command to completely fail when the current filename contained such
characters.

This commit fixes that by replacing all such characters with underscores ('_').

fixes #140